+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/constants/BigDecimalConstants.java
@@ -7,5 +7,6 @@ import java.math.BigDecimal;
 @UtilityClass
 public class BigDecimalConstants {
     public static final BigDecimal MINUS_ONE = BigDecimal.valueOf(-1);
+    public static final BigDecimal TWO = BigDecimal.valueOf(2L);
     public static final BigDecimal ONE_HUNDRED = BigDecimal.valueOf(100);
 }